refactor: create a `faucet_control` module
Rather than having a `faucet` module within `node_control`, I thought it made sense to also have a
`faucet_control` module, and I've organised it the same as `node_control`.
